+++ docs/changelog.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Web-socket data size validation
 ### Changed
  - Change single logger to a logging hierarchy
+ - Decompose models
+ - Optimize ScTemplate, ScTemplateResult and change list to tuple in params
 ### Deprecated
  - ScModule, ScAgent and ScKeynode implementations moved to [py-sc-kpm](https://github.com/ostis-ai/py-sc-kpm)